requirements.txt
Flask 
python-pptx
openai
tqdm
beautifulsoup4
python-dotenv